Fix jar hashing and comparison of the Updater on Windows
By comparing a full hash of the jar we are also hashing the build timestamp,
meaning that the check always resulted in that the jar is different even
though the content didn't change.

By hashing all entries in the jar we avoid this, because they don't
include any unique build timestamps.